+++ common/src/main/java/nl/enjarai/doabarrelroll/util/MixinHooks.java
@@ -1,0 +1,10 @@
+package nl.enjarai.doabarrelroll.util;
+
+public class MixinHooks {
+    /**
+     * Used for both TRIPLE_JUMP and HYBRID activation behaviours.
+     */
+    public static boolean secondJump = false;
+    public static boolean thirdJump = false;
+    public static boolean wasJumping = false;
+}