+++ fabric/src/main/java/nl/enjarai/doabarrelroll/compat/cameraoverhaul/CameraOverhaulPlugin.java
@@ -1,0 +1,12 @@
+package nl.enjarai.doabarrelroll.compat.cameraoverhaul;
+
+import nl.enjarai.doabarrelroll.compat.CompatMixinPlugin;
+
+import java.util.Set;
+
+public class CameraOverhaulPlugin implements CompatMixinPlugin {
+    @Override
+    public Set<String> getRequiredMods() {
+        return Set.of("cameraoverhaul");
+    }
+}